+++ CHANGELOG.md
@@ -4,11 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## 1.x.x - 2021-04-09
 
 ### Added
+- Added Montserrat as default font
 - Added `Forbidden` and `NotFound` error pages
 - Automatic import of error pages
+
+### Removed
+- Removed `#q-app` background-color 
 
 ## 1.7.0 - 2021-04-13