+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Added Montserrat as default font
+- Added `Forbidden` and `NotFound` error pages
+- Automatic import of error pages
 
 ### Removed
 - Removed `#q-app` background-color