+++ CHANGELOG.md
@@ -4,12 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## 1.x.x - 2021-04-09
+
+### Added
+- Added Montserrat as default font
+- Added `Forbidden` and `NotFound` error pages
+- Automatic import of error pages
 
 ### Changed
 - Start using `expanded-icon` on `QasAppMenu`
 
-## Fixed
+### Removed
+- Removed `#q-app` background-color
+
+### Fixed
 - Prefix all icons with `o_` to force Material Icons Outlined
 
 ## 1.7.0 - 2021-04-13