+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.x.x - 2021-04-20
+
+### Added
+- Added component `qas-tip` in `ui/src/components`
+
 ## 1.x.x - 2021-04-09
 
 ### Added